+++ src/harmonist/formats/ogg.py
@@ -39,6 +39,7 @@ read_duration_ms = _impl.read_duration_ms
 write_tags = _impl.write_tags
 write_owned = _impl.write_owned
 read_owned = _impl.read_owned
+has_superseded_tags = _impl.has_superseded_tags
 write_cover = _impl.write_cover
 read_cover = _impl.read_cover